Extract collection comprehension tests to own file
