+++ cpg-language-python/src/test/kotlin/de/fraunhofer/aisec/cpg/frontends/python/CollectionComprehensionPython3Test.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024, Fraunhofer AISEC. All rights reserved.
+ * Copyright (c) 2025, Fraunhofer AISEC. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -25,90 +25,159 @@
  */
 package de.fraunhofer.aisec.cpg.frontends.python
 
-import de.fraunhofer.aisec.cpg.graph.*
+import de.fraunhofer.aisec.cpg.TranslationResult
 import de.fraunhofer.aisec.cpg.graph.declarations.FunctionDeclaration
 import de.fraunhofer.aisec.cpg.graph.declarations.ParameterDeclaration
 import de.fraunhofer.aisec.cpg.graph.declarations.VariableDeclaration
-import de.fraunhofer.aisec.cpg.graph.scopes.FunctionScope
-import de.fraunhofer.aisec.cpg.graph.statements.expressions.*
-import de.fraunhofer.aisec.cpg.test.*
+import de.fraunhofer.aisec.cpg.graph.functions
+import de.fraunhofer.aisec.cpg.graph.get
+import de.fraunhofer.aisec.cpg.graph.invoke
+import de.fraunhofer.aisec.cpg.graph.refs
+import de.fraunhofer.aisec.cpg.graph.scopes.LocalScope
+import de.fraunhofer.aisec.cpg.graph.statements
+import de.fraunhofer.aisec.cpg.graph.statements.expressions.AssignExpression
+import de.fraunhofer.aisec.cpg.graph.statements.expressions.BinaryOperator
+import de.fraunhofer.aisec.cpg.graph.statements.expressions.Block
+import de.fraunhofer.aisec.cpg.graph.statements.expressions.CallExpression
+import de.fraunhofer.aisec.cpg.graph.statements.expressions.CollectionComprehension
+import de.fraunhofer.aisec.cpg.graph.statements.expressions.InitializerListExpression
+import de.fraunhofer.aisec.cpg.graph.statements.expressions.KeyValueExpression
+import de.fraunhofer.aisec.cpg.graph.statements.expressions.Reference
+import de.fraunhofer.aisec.cpg.graph.variables
+import de.fraunhofer.aisec.cpg.test.analyze
+import de.fraunhofer.aisec.cpg.test.assertLocalName
+import de.fraunhofer.aisec.cpg.test.assertNotRefersTo
+import de.fraunhofer.aisec.cpg.test.assertRefersTo
 import java.nio.file.Path
-import kotlin.test.*
+import kotlin.test.assertEquals
+import kotlin.test.assertIs
+import kotlin.test.assertNotNull
+import kotlin.test.assertNull
+import org.junit.jupiter.api.BeforeAll
+import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.TestInstance
 
-class ExpressionHandlerPython2Test {
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
+class CollectionComprehensionPython3Test {
+    private lateinit var result: TranslationResult
 
-    @Test
-    fun testComprehensionExpressionTuplePython2() {
+    @BeforeAll
+    fun setup() {
         val topLevel = Path.of("src", "test", "resources", "python")
-        val result =
+        result =
             analyze(listOf(topLevel.resolve("comprehension.py").toFile()), topLevel, true) {
                 it.registerLanguage<PythonLanguage>()
                 it.symbols(
                     mapOf(
-                        "PYTHON_VERSION_MAJOR" to "2",
+                        "PYTHON_VERSION_MAJOR" to "3",
                         "PYTHON_VERSION_MINOR" to "0",
                         "PYTHON_VERSION_MICRO" to "0",
                     )
                 )
             }
         assertNotNull(result)
-
-        val tupleComp = result.functions["tuple_comp"]
-        assertNotNull(tupleComp)
-
-        val body = tupleComp.body
-        assertIs<Block>(body)
-        val tupleAsVariableAssignment = body.statements[0]
-        assertIs<AssignExpression>(tupleAsVariableAssignment)
-        val tupleAsVariable = tupleAsVariableAssignment.rhs[0]
-        assertIs<CollectionComprehension>(tupleAsVariable)
-        val barCall = tupleAsVariable.statement
-        assertIs<CallExpression>(barCall)
-        assertLocalName("bar", barCall)
-        val argK = barCall.arguments[0]
-        assertIs<Reference>(argK)
-        assertLocalName("k", argK)
-        val argV = barCall.arguments[1]
-        assertIs<Reference>(argV)
-        assertLocalName("v", argV)
-        assertEquals(1, tupleAsVariable.comprehensionExpressions.size)
-        val initializerListExpression = tupleAsVariable.comprehensionExpressions[0].variable
-        assertIs<InitializerListExpression>(initializerListExpression)
-        val variableK = initializerListExpression.initializers[0]
-        assertIs<Reference>(variableK)
-        assertLocalName("k", variableK)
-        val variableV = initializerListExpression.initializers[1]
-        assertIs<Reference>(variableV)
-        assertLocalName("v", variableV)
-
-        // Check that the declarations exist for the variables k and v
-        val declK = variableK.refersTo
-        assertIs<VariableDeclaration>(declK)
-        assertIs<FunctionScope>(declK.scope)
-        assertEquals(tupleComp, declK.scope?.astNode)
-        assertRefersTo(argK, declK)
-        val declV = variableV.refersTo
-        assertIs<VariableDeclaration>(declV)
-        assertIs<FunctionScope>(declV.scope)
-        assertEquals(tupleComp, declV.scope?.astNode)
-        assertRefersTo(argV, declV)
     }
 
     @Test
-    fun testListComprehensionsPython2() {
-        val topLevel = Path.of("src", "test", "resources", "python")
-        val result =
-            analyze(listOf(topLevel.resolve("comprehension.py").toFile()), topLevel, true) {
-                it.registerLanguage<PythonLanguage>()
-                it.symbols(
-                    mapOf(
-                        "PYTHON_VERSION_MAJOR" to "2",
-                        "PYTHON_VERSION_MINOR" to "0",
-                        "PYTHON_VERSION_MICRO" to "0",
-                    )
-                )
-            }
-        assertNotNull(result)
+    fun testComprehensionExpressionTuplePython3() {
+        // Get the function tuple_comp
+        val tupleComp = result.functions["tuple_comp"]
+        assertNotNull(tupleComp, "There was no function \"tuple_comp\"")
+
+        // Get the body
+        val body = tupleComp.body
+        assertIs<Block>(body, "The body of \"tuple_comp\" must be a Block.")
+        // The first statement is an assigment of a list comprehension with an if to a variable "a"
+        val tupleAsVariableAssignment = body.statements[0]
+        assertIs<AssignExpression>(
+            tupleAsVariableAssignment,
+            "The statement is an AssignExpression",
+        )
+        val tupleAsVariable = tupleAsVariableAssignment.rhs[0]
+        assertIs<CollectionComprehension>(
+            tupleAsVariable,
+            "The right hand side must be a CollectionComprehension representing python's list comprehension \"[bar(k, v) for (k, v) in x]\".",
+        )
+        val barCall = tupleAsVariable.statement
+        assertIs<CallExpression>(
+            barCall,
+            "The statement inside the list comprehension is a call to bar with arguments k and v",
+        )
+        assertLocalName("bar", barCall, "The CallExpression calls bar()")
+        val argK = barCall.arguments[0]
+        assertIs<Reference>(argK, "The first argument of bar() is a reference k")
+        assertLocalName("k", argK, "The first argument of bar() is a reference k")
+        val argV = barCall.arguments[1]
+        assertIs<Reference>(argV, "The second argument of bar() is a reference v")
+        assertLocalName("v", argV, "The second argument of bar() is a reference v")
+        assertEquals(
+            1,
+            tupleAsVariable.comprehensionExpressions.size,
+            "There is a single comprehension expression (\"for (k, v) in x\")",
+        )
+        val initializerListExpression = tupleAsVariable.comprehensionExpressions[0].variable
+        assertIs<InitializerListExpression>(
+            initializerListExpression,
+            "The variable is actually tuple which is represented as an InitializerListExpression in the CPG",
+        )
+        val variableK = initializerListExpression.initializers[0]
+        assertIs<Reference>(
+            variableK,
+            "The first element in the tuple is a variable reference \"k\"",
+        )
+        assertLocalName(
+            "k",
+            variableK,
+            "The first element in the tuple is a variable reference \"k\"",
+        )
+        val variableV = initializerListExpression.initializers[1]
+        assertIs<Reference>(
+            variableV,
+            "The second element in the tuple is a variable reference \"v\"",
+        )
+        assertLocalName(
+            "v",
+            variableV,
+            "The second element in the tuple is a variable reference \"V\"",
+        )
+
+        // Check that the declarations exist for the variables k and v
+        val declK = variableK.refersTo
+        assertIs<VariableDeclaration>(declK, "The refersTo should be a VariableDeclaration")
+        assertIs<LocalScope>(
+            declK.scope,
+            "The scope of the variable is the local scope belonging to the list comprehension. In particular, it is not the FunctionScope.",
+        )
+        assertEquals(
+            tupleAsVariable,
+            declK.scope?.astNode,
+            "The scope of the variable is the local scope belonging to the list comprehension. In particular, it is not the FunctionScope.",
+        )
+        assertRefersTo(
+            argK,
+            declK,
+            "The argument k of the call also refers to the variable k declared in the comprehension expression.",
+        )
+        val declV = variableV.refersTo
+        assertIs<VariableDeclaration>(declV, "The refersTo should be a VariableDeclaration")
+        assertIs<LocalScope>(
+            declV.scope,
+            "The scope of the variable is the local scope belonging to the list comprehension. In particular, it is not the FunctionScope.",
+        )
+        assertEquals(
+            tupleAsVariable,
+            declV.scope?.astNode,
+            "The scope of the variable is the local scope belonging to the list comprehension. In particular, it is not the FunctionScope.",
+        )
+        assertRefersTo(
+            argV,
+            declV,
+            "The argument v of the call also refers to the variable v declared in the comprehension expression.",
+        )
+    }
+
+    @Test
+    fun testListComprehensionsPython3() {
         val listComp = result.functions["list_comp"]
         assertNotNull(listComp)
         val paramX = listComp.parameters[0]
@@ -131,7 +200,7 @@ class ExpressionHandlerPython2Test {
         assertLocalName("i", variableI)
         val declI = variableI.refersTo
         assertIs<VariableDeclaration>(declI)
-        assertEquals(listComp, declI.scope?.astNode)
+        assertEquals(singleWithIf, declI.scope?.astNode)
         val iterableX = singleWithIf.comprehensionExpressions[0].iterable
         assertIs<Reference>(iterableX)
         assertLocalName("x", iterableX)
@@ -146,7 +215,7 @@ class ExpressionHandlerPython2Test {
         val outsideI = fooIOutside.arguments[0]
         assertIs<Reference>(outsideI)
         assertLocalName("i", outsideI)
-        assertRefersTo(outsideI, declI)
+        assertNotRefersTo(outsideI, declI)
 
         val singleWithoutIfAssignment = body.statements[1]
         assertIs<AssignExpression>(singleWithoutIfAssignment)
@@ -182,20 +251,7 @@ class ExpressionHandlerPython2Test {
     }
 
     @Test
-    fun testSetComprehensionsPython2() {
-        val topLevel = Path.of("src", "test", "resources", "python")
-        val result =
-            analyze(listOf(topLevel.resolve("comprehension.py").toFile()), topLevel, true) {
-                it.registerLanguage<PythonLanguage>()
-                it.symbols(
-                    mapOf(
-                        "PYTHON_VERSION_MAJOR" to "2",
-                        "PYTHON_VERSION_MINOR" to "0",
-                        "PYTHON_VERSION_MICRO" to "0",
-                    )
-                )
-            }
-        assertNotNull(result)
+    fun testSetComprehensionsPython3() {
         val listComp = result.functions["set_comp"]
         assertNotNull(listComp)
 
@@ -247,20 +303,7 @@ class ExpressionHandlerPython2Test {
     }
 
     @Test
-    fun testDictComprehensionsPython2() {
-        val topLevel = Path.of("src", "test", "resources", "python")
-        val result =
-            analyze(listOf(topLevel.resolve("comprehension.py").toFile()), topLevel, true) {
-                it.registerLanguage<PythonLanguage>()
-                it.symbols(
-                    mapOf(
-                        "PYTHON_VERSION_MAJOR" to "2",
-                        "PYTHON_VERSION_MINOR" to "0",
-                        "PYTHON_VERSION_MICRO" to "0",
-                    )
-                )
-            }
-        assertNotNull(result)
+    fun testDictComprehensionsPython3() {
         val listComp = result.functions["dict_comp"]
         assertNotNull(listComp)
 
@@ -328,20 +371,7 @@ class ExpressionHandlerPython2Test {
     }
 
     @Test
-    fun testGeneratorExprPython2() {
-        val topLevel = Path.of("src", "test", "resources", "python")
-        val result =
-            analyze(listOf(topLevel.resolve("comprehension.py").toFile()), topLevel, true) {
-                it.registerLanguage<PythonLanguage>()
-                it.symbols(
-                    mapOf(
-                        "PYTHON_VERSION_MAJOR" to "2",
-                        "PYTHON_VERSION_MINOR" to "0",
-                        "PYTHON_VERSION_MICRO" to "0",
-                    )
-                )
-            }
-        assertNotNull(result)
+    fun testGeneratorExprPython3() {
         val listComp = result.functions["generator"]
         assertNotNull(listComp)
 
@@ -377,20 +407,7 @@ class ExpressionHandlerPython2Test {
      * This test ensures that variables in a comprehension do not bind to the outer scope. See
      * [testCompBindingAssignExpr] for exceptions.
      */
-    fun testCompBindingPython2() {
-        val topLevel = Path.of("src", "test", "resources", "python")
-        val result =
-            analyze(listOf(topLevel.resolve("comprehension.py").toFile()), topLevel, true) {
-                it.registerLanguage<PythonLanguage>()
-                it.symbols(
-                    mapOf(
-                        "PYTHON_VERSION_MAJOR" to "2",
-                        "PYTHON_VERSION_MINOR" to "0",
-                        "PYTHON_VERSION_MICRO" to "0",
-                    )
-                )
-            }
-        assertNotNull(result)
+    fun testCompBindingPython3() {
         val compBindingFunc = result.functions["comp_binding"]
         assertIs<FunctionDeclaration>(compBindingFunc)
 
@@ -398,22 +415,22 @@ class ExpressionHandlerPython2Test {
         assertIs<VariableDeclaration>(xDecl)
 
         assertEquals(
-            1,
+            5,
             compBindingFunc.variables.size,
-            "Expected one variable where all other \"x\" refer to",
+            "Expected five variables. One for the \"outside\" x and one for each of the four comprehensions.",
         )
 
         assertEquals(
-            11,
+            2,
             xDecl.usages.size,
-            "Expected eleven usages: one for the initial assignment and one for the usage in \"print(x)\" and nine inside the list/set/dict comprehensions.",
+            "Expected two usages: one for the initial assignment and one for the usage in \"print(x)\".",
         )
 
         val comprehensions =
             compBindingFunc.body.statements.filterIsInstance<CollectionComprehension>()
         assertEquals(4, comprehensions.size, "Expected to find four comprehensions.")
 
-        comprehensions.forEach { it.refs("x").forEach { ref -> assertRefersTo(ref, xDecl) } }
+        comprehensions.forEach { it.refs("x").forEach { ref -> assertNotRefersTo(ref, xDecl) } }
     }
 
     @Test
@@ -421,20 +438,7 @@ class ExpressionHandlerPython2Test {
      * This test ensures that variables in a comprehension do not bind to the outer scope if they
      * are used in an `AssignExpr`. See https://peps.python.org/pep-0572/#scope-of-the-target
      */
-    fun testCompBindingAssignExprPython2() {
-        val topLevel = Path.of("src", "test", "resources", "python")
-        val result =
-            analyze(listOf(topLevel.resolve("comprehension.py").toFile()), topLevel, true) {
-                it.registerLanguage<PythonLanguage>()
-                it.symbols(
-                    mapOf(
-                        "PYTHON_VERSION_MAJOR" to "2",
-                        "PYTHON_VERSION_MINOR" to "0",
-                        "PYTHON_VERSION_MICRO" to "0",
-                    )
-                )
-            }
-        assertNotNull(result)
+    fun testCompBindingAssignExprPython3() {
 
         val compBindingAssignFunc = result.functions["comp_binding_assign_expr"]
         assertIs<FunctionDeclaration>(compBindingAssignFunc)
@@ -443,9 +447,9 @@ class ExpressionHandlerPython2Test {
         assertIs<VariableDeclaration>(xDecl)
 
         assertEquals(
-            1,
+            2,
             compBindingAssignFunc.variables.size,
-            "Expected two variables. One for the \"outside\" x.",
+            "Expected two variables. One for the \"outside\" x and one for the \"temp\" inside the comprehension.",
         )
 
         assertEquals(
@@ -460,9 +464,53 @@ class ExpressionHandlerPython2Test {
         val xRef = comprehension.refs("x").singleOrNull()
         assertNotNull(xRef)
         assertRefersTo(xRef, xDecl)
+    }
 
-        // TODO: test for other types of comprehensions
-        // TODO: test nested comprehensions -> AssignExpression binds to
-        // containing scope
+    @Test
+    /**
+     * This test ensures that variables in a comprehension do not bind to the outer scope if they
+     * are used in an `AssignExpr`. See https://peps.python.org/pep-0572/#scope-of-the-target
+     */
+    fun testCompBindingAssignExprNestedPython3() {
+
+        val compBindingAssignFunc = result.functions["comp_binding_assign_expr_nested"]
+        assertIs<FunctionDeclaration>(compBindingAssignFunc)
+
+        val xDecl = compBindingAssignFunc.variables.firstOrNull()
+        assertIs<VariableDeclaration>(xDecl)
+
+        assertEquals(
+            3,
+            compBindingAssignFunc.variables.size,
+            "Expected two variables. One for the \"outside\" x, one for the \"temp\" inside the comprehension and one for the \"a\" inside the comprehension.",
+        )
+
+        assertEquals(
+            3,
+            xDecl.usages.size,
+            "Expected three usages: one for the initial assignment, one for the comprehension and one for the usage in \"print(x)\".",
+        )
+        val body = compBindingAssignFunc.body
+        assertIs<Block>(body, "The body of a function must be a Block.")
+        val outerComprehension = body.statements.singleOrNull { it is CollectionComprehension }
+        assertIs<CollectionComprehension>(
+            outerComprehension,
+            "There must be exactly one CollectionComprehension (the list comprehension) in the statement of the body. Note: The inner collection comprehension would be reached by the extension function Node.statements which does not apply here.",
+        )
+        val innerComprehension = outerComprehension.statement
+        assertIs<CollectionComprehension>(
+            innerComprehension,
+            "The inner comprehension is the statement of the outer list comprehension",
+        )
+        val xRef = innerComprehension.refs("x").singleOrNull()
+        assertNotNull(
+            xRef,
+            "There is only one usage of \"x\" which is inside the inner comprehension's statement.",
+        )
+        assertRefersTo(
+            xRef,
+            xDecl,
+            "The reference of \"x\" inside the inner comprehension's statement refers to the variable declared outside the comprehensions.",
+        )
     }
 }
